first instruction moveq

+++ src/cpu/instructions.go
@@ -6,7 +6,7 @@ func (cpu *M68k) init68000InstructionSet() {
 	registerMove(cpu)
 }
 
-func addInstruction(cpu *M68k, opcode uint16, instruction Instruction) {
+func addInstruction(cpu *M68k, opcode int, instruction Instruction) {
 	if cpu.instructions[opcode] != nil {
 		panic(fmt.Errorf("Opcode $%04x already set"))
 	}
